refactoring: 0.5 in estimations
git-svn-id: https://servisto.de/svn/share/trunk/scrum@3201 15bc23fa-3125-0410-a1f3-bdc40dbf1900

+++ src/main/java/scrum/client/estimation/PlanningPokerCardWidget.java
@@ -9,11 +9,11 @@ import com.google.gwt.user.client.ui.Widget;
 
 public class PlanningPokerCardWidget extends AScrumWidget {
 
-	private int value;
+	private float value;
 	private ClickHandler clickHandler;
 	private boolean visible;
 
-	public PlanningPokerCardWidget(int value, boolean visible, ClickHandler clickHandler) {
+	public PlanningPokerCardWidget(float value, boolean visible, ClickHandler clickHandler) {
 		super();
 		this.value = value;
 		this.visible = visible;
