+++ src/main/java/scrum/client/project/EstimationBarFactory.java
@@ -7,17 +7,17 @@ public class EstimationBarFactory {
 
 	public static void createEstimationBars(List<Requirement> requirements, Integer velocity) {
 		int sprintOffset = 0;
-		int remainingWorkInSprint = velocity == null ? 0 : velocity;
-		List<Integer> workPerSprint;
+		float remainingWorkInSprint = velocity == null ? 0 : velocity;
+		List<Float> workPerSprint;
 		for (Requirement requirement : requirements) {
-			workPerSprint = new ArrayList<Integer>();
-			Integer estimatedWork = requirement.getEstimatedWork();
+			workPerSprint = new ArrayList<Float>();
+			Float estimatedWork = requirement.getEstimatedWork();
 			if (velocity == null || estimatedWork == null) {
 				if (estimatedWork != null) workPerSprint.add(estimatedWork);
 				requirement.setEstimationBar(new EstimationBar(sprintOffset, workPerSprint));
 				continue;
 			}
-			int work = estimatedWork;
+			float work = estimatedWork;
 			while (work > remainingWorkInSprint) {
 				workPerSprint.add(remainingWorkInSprint);
 				work -= remainingWorkInSprint;
